+++ src/Group.js
@@ -24,7 +24,7 @@
         _validateAdd: function(child) {
             var type = child.getType();
             if (type !== 'Group' && type !== 'Shape') {
-                Konva.Util.error('You may only add groups and shapes to groups.');
+                Konva.Util.throw('You may only add groups and shapes to groups.');
             }
         }
     });